[Keymap] Move my custom functions and keymaps to userspace (#10502)
* Move my custom functions and keymaps to userspace to be able to add the STM32F411 variant of my keyboard with mostly the same functionality

* Move keymaps to community layouts

* Remove keymaps from userspace readme.md
