+++ keyboards/handwired/riblee_f401/riblee_f401.h
@@ -1,5 +1,5 @@
 /*
-Copyright 2020 QMK Community, Daniel Reibl <janos.daniel.reibl@protonmail.com>
+Copyright 2020 QMK Community, Janos Daniel Reibl <janos.daniel.reibl@protonmail.com> @riblee
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
