+++ keyboards/handwired/riblee_f401/rules.mk
@@ -22,3 +22,5 @@ BACKLIGHT_DRIVER = software
 RGBLIGHT_ENABLE = no        # Enable keyboard RGB underglow
 BLUETOOTH_ENABLE = no       # Enable Bluetooth
 AUDIO_ENABLE = no           # Audio output
+
+LAYOUTS = ortho_5x12
